+++ src/lib/ltree/ltree-sort.test.ts
@@ -11,8 +11,11 @@ describe('LTree Sorting', () => {
 			undefined, // levelMember
 			undefined, // hasChildrenMember
 			undefined, // isExpandedMember
+			undefined, // getIsExpandedCallback
 			undefined, // isSelectableMember
+			undefined, // getIsSelectableCallback
 			undefined, // isSelectedMember
+			undefined, // getIsSelectedCallback
 			undefined, // isDraggableMember
 			undefined, // getIsDraggableCallback
 			undefined, // isDropAllowedMember